refactor: services/hooks를 도메인별 서브폴더로 분리 (#45) - hooks/ 하위에 auth, theme, design, user, common 폴더 생성 - 각 훅을 도메인에 따라 분류하여 이동 - useBoardWriteForm, useIntersectionObserver, useOutsideClick은   도메인 독립적이므로 common/으로 분리 - 이동에 따른 내부 상대경로 및 외부 import 경로 전체 업데이트
