+++ src/components/community/CommentModal.tsx
@@ -3,7 +3,7 @@ import HeartIcon from "../icons/community-detail/heart.svg?react";
 import type { ICommentRaw } from "../../types/community/theme";
 import Text from "../common/Text";
 import { useAuthStore } from "../../stores/authStore";
-import { useOutsideClick } from "../../services/hooks/useOutsideClick";
+import { useOutsideClick } from "../../services/hooks/common/useOutsideClick";
 import { usePostMutation } from "../../services/api/useApi";
 import { ThemeService } from "../../services/api/ThemeService";
 import { useQueryClient } from "@tanstack/react-query";
