+++ src/components/community/design/DesignDetailCard.tsx
@@ -3,11 +3,11 @@ import { createPortal } from "react-dom";
 import { useNavigate } from "react-router-dom";
 import type { IDesignBoardDetail } from "../../../types/community/design";
 import Text from "../../common/Text";
-import { useOutsideClick } from "../../../services/hooks/useOutsideClick";
-import { useComments } from "../../../services/hooks/useComments";
-import { useCommentActions } from "../../../services/hooks/useCommentActions";
-import { usePrefer } from "../../../services/hooks/usePrefer";
-import { useDeleteDesignPost } from "../../../services/hooks/useDeleteDesignPost";
+import { useOutsideClick } from "../../../services/hooks/common/useOutsideClick";
+import { useComments } from "../../../services/hooks/theme/useComments";
+import { useCommentActions } from "../../../services/hooks/theme/useCommentActions";
+import { usePrefer } from "../../../services/hooks/theme/usePrefer";
+import { useDeleteDesignPost } from "../../../services/hooks/design/useDeleteDesignPost";
 import { useAuthStore } from "../../../stores/authStore";
 import CommentModal from "../CommentModal";
 import Confirm from "../../common/Confirm";
